SDCS-X/vIT-Management/requirements.txt
# Requirements
fastapi==0.101.0
uvicorn==0.27.1
paho-mqtt==1.6.1
docker==5.0.3
influxdb-client==1.40.0
jinja2==3.1.2

